+++ src/pages/iv-ava-by-site/index.js
@@ -8,7 +8,7 @@ import { Fixed } from 'src/components/Shared/Layouts/Fixed'
 import { VertLayout } from 'src/components/Shared/Layouts/VertLayout'
 import { Grow } from 'src/components/Shared/Layouts/Grow'
 import RPBGridToolbar from 'src/components/Shared/RPBGridToolbar'
-import SerialForm from './forms/SerialForm'
+import SerialTable from './forms/SerialTable'
 import serialIcon from '../../../public/images/TableIcons/imgSerials.png'
 import lotIcon from '../../../public/images/TableIcons/lot.png'
 import { Box, IconButton } from '@mui/material'
@@ -153,7 +153,7 @@ const AvailabilitiesBySite = () => {
 
   function openSerialForm(itemId, siteId) {
     stack({
-      Component: SerialForm,
+      Component: SerialTable,
       props: {
         labels,
         itemId,
