+++ src/repositories/InventoryRepository.js
@@ -264,7 +264,6 @@ export const InventoryRepository = {
   },
   Serial: {
     qry: service + 'qrySRL',
-    snapshot: service + 'snapshotSRL',
     get2: service + 'get2SRL',
     get: service + 'getSRL'
   },
